pyproject.toml
[project]
name = "ai-digital-twin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.30.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "gradio>=6.0,<7",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
